Move project config discovery to seperate file
